fix(gateway): desktop stream dropped-frame handling, shared guardian-stream tests, comment trims
